Adds import guard to avoid circular dependency
We need to clean this up but for now this will work OK. We have a if
TYPE_CHECKING in graph_types due to a strange issue that only occurs
during build for docs. While this indicates a code smell, this also
suggests issues around the way the path is set up for documentation.